# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
}

export interface CurrentUser {
  user: User | null;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface AnalyzeInput {
  /**
     * @minLength 1
     * @maxLength 2048
     */
  url: string;
}

export type ScanFeatureStatus = typeof ScanFeatureStatus[keyof typeof ScanFeatureStatus];


export const ScanFeatureStatus = {
  positive: 'positive',
  caution: 'caution',
  danger: 'danger',
} as const;

export interface ScanFeature {
  label: string;
  value: string;
  status: ScanFeatureStatus;
  detail: string;
}

export type ScanClassification = typeof ScanClassification[keyof typeof ScanClassification];


export const ScanClassification = {
  safe: 'safe',
  suspicious: 'suspicious',
  'high-risk': 'high-risk',
} as const;

export interface Scan {
  id: string;
  url: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  riskScore: number;
  classification: ScanClassification;
  prediction: string;
  createdAt: string;
  riskLevel: string;
  features: ScanFeature[];
  positiveIndicators: string[];
  riskIndicators: string[];
  recommendation: string;
}

export type DashboardActivityItem = {
  label: string;
  scans: number;
  risk: number;
};

export interface Dashboard {
  totalScans: number;
  safeCount: number;
  suspiciousCount: number;
  highRiskCount: number;
  averageRiskScore: number;
  recentScans: Scan[];
  activity: DashboardActivityItem[];
  modelStatus: string;
}

export interface QuizResultInput {
  /** @minimum 0 */
  score: number;
  /** @minimum 1 */
  totalQuestions: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  percentage: number;
}

export type QuizResult = QuizResultInput & {
  id: string;
  createdAt: string;
};

export type Profile = User & {
  totalScans: number;
  safeScans: number;
  suspiciousScans: number;
  highRiskScans: number;
};

export type ListScansParams = {
q?: string;
classification?: ListScansClassification;
};

export type ListScansClassification = typeof ListScansClassification[keyof typeof ListScansClassification];


export const ListScansClassification = {
  all: 'all',
  safe: 'safe',
  suspicious: 'suspicious',
  'high-risk': 'high-risk',
} as const;

export type DeleteScanParams = {
id: string;
};