# lib/api-zod/src/generated/types/analyzeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeInput {
  /**
     * @minLength 1
     * @maxLength 2048
     */
  url: string;
}