# lib/api-zod/src/generated/types/currentUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface CurrentUser {
  user: User | null;
}