# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardActivityItem } from './dashboardActivityItem';
import type { Scan } from './scan';

export interface Dashboard {
  totalScans: number;
  safeCount: number;
  suspiciousCount: number;
  highRiskCount: number;
  averageRiskScore: number;
  recentScans: Scan[];
  activity: DashboardActivityItem[];
  modelStatus: string;
}