# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export type Profile = User & {
  totalScans: number;
  safeScans: number;
  suspiciousScans: number;
  highRiskScans: number;
};