# lib/api-zod/src/generated/types/quizResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizResultInput } from './quizResultInput';

export type QuizResult = QuizResultInput & {
  id: string;
  createdAt: string;
};