# lib/api-zod/src/generated/types/quizResultInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuizResultInput {
  /** @minimum 0 */
  score: number;
  /** @minimum 1 */
  totalQuestions: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  percentage: number;
}