# lib/api-zod/src/generated/types/scan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanClassification } from './scanClassification';
import type { ScanFeature } from './scanFeature';

export interface Scan {
  id: string;
  url: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  riskScore: number;
  classification: ScanClassification;
  prediction: string;
  createdAt: string;
  riskLevel: string;
  features: ScanFeature[];
  positiveIndicators: string[];
  riskIndicators: string[];
  recommendation: string;
}