# lib/api-zod/src/generated/types/scanFeature.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanFeatureStatus } from './scanFeatureStatus';

export interface ScanFeature {
  label: string;
  value: string;
  status: ScanFeatureStatus;
  detail: string;
}